requirements.txt
Django
djangorestframework
django-ckeditor
redis
gunicorn